+++ controlers/contacts.js
@@ -1,14 +1,6 @@
-const Joi = require("Joi");
-
 const contacts = require("../models/contacts");
 
 const { HttpError, ctrlWrapper } = require("../helpers");
-
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.number().required(),
-});
 
 const listContacts = async (req, res) => {
   const allContacts = await contacts.listContacts();
@@ -25,10 +17,6 @@ const getContactById = async (req, res) => {
 };
 
 const addContact = async (req, res) => {
-  const { error } = addSchema.validate(req.body);
-  if (error) {
-    throw HttpError(400, "missing required name field");
-  }
   const newContact = await contacts.addContact(req.body);
   res.status(201).json(newContact);
 };
@@ -43,10 +31,6 @@ const removeContact = async (req, res) => {
 };
 
 const updateContact = async (req, res) => {
-  const { error } = addSchema.validate(req.body);
-  if (error) {
-    throw HttpError(400, "missing fields");
-  }
   const { contactId } = req.params;
   const updateContact = await contacts.updateContact(contactId, req.body);
   if (!updateContact) {
