+++ routes/api/contacts.js
@@ -2,16 +2,20 @@ const express = require("express");
 
 const ctrl = require("../../controlers/contacts");
 
+const { validateBody } = require("../../middlewares");
+
+const schemas = require("../../schemas/contacts");
+
 const router = express.Router();
 
 router.get("/", ctrl.listContacts);
 
 router.get("/:contactId", ctrl.getContactById);
 
-router.post("/", ctrl.addContact);
+router.post("/", validateBody(schemas.addSchema), ctrl.addContact);
 
 router.delete("/:contactId", ctrl.removeContact);
 
-router.put("/:contactId", ctrl.updateContact);
+router.put("/:contactId", validateBody(schemas.addSchema), ctrl.updateContact);
 
 module.exports = router;
